MWPW-165727-moved tests
